Dalamud SDK 15. Also add temp placeholders for CkRaii while other updates get resolved.

+++ DalamudServices/PlayerContent.cs
@@ -13,7 +13,7 @@ namespace CkCommons;
 public static class PlayerContent
 {
     private static Dictionary<uint, string> Cache = [];
-    public static ushort TerritoryID => Svc.ClientState.TerritoryType;
+    public static uint TerritoryID => Svc.ClientState.TerritoryType;
 
     public static string TerritoryName
     {
